添加节点类型、标签等

+++ xm-core/src/main/java/com/xm/core/entity/XmMenuState.java
@@ -1,16 +1,15 @@
-package com.xm.core.entity;
+package  com.xm.core.entity;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-
-import java.math.BigDecimal;
 import java.util.Date;
+import java.math.BigDecimal;
 
 /**
- * 组织 com.qqkj  顶级模块 xm 大模块 core  小模块 <br> 
+ * 组织 com  顶级模块 xm 大模块 core  小模块 <br> 
  * 实体 XmMenuState所有属性名: <br>
  *	id,menuId,planStartTime,planEndTime,actStartTime,actEndTime,planWorkload,actWorkload,planCostAmount,actCostAmount,finishRate,demandRate,designRate,devRate,uatRate,sitRate,onlineStatus,onlineTime,planStatus,chargeUserid,chargeUsername,menuStatus,ctime,ltime,cuserid,cusername,calcTime,menuName,planWorkhours,planWorkerCnt,closedBugs,activeBugs,confirmedBugs,resolvedBugs,productId,productName,testCases,execCases,designCases,finishCases,projectCnt,iterationCnt,taskCnt,finishTaskCnt,bizDate,bugCnt;<br>
- * 表 XM.xm_menu_state 功能状态表,无需前端维护，所有数据由汇总统计得出的所有字段名: <br>
+ * 表 xm_menu_state 功能状态表,无需前端维护，所有数据由汇总统计得出的所有字段名: <br>
  *	id,menu_id,plan_start_time,plan_end_time,act_start_time,act_end_time,plan_workload,act_workload,plan_cost_amount,act_cost_amount,finish_rate,demand_rate,design_rate,dev_rate,uat_rate,sit_rate,online_status,online_time,plan_status,charge_userid,charge_username,menu_status,ctime,ltime,cuserid,cusername,calc_time,menu_name,plan_workhours,plan_worker_cnt,closed_bugs,active_bugs,confirmed_bugs,resolved_bugs,product_id,product_name,test_cases,exec_cases,design_cases,finish_cases,project_cnt,iteration_cnt,task_cnt,finish_task_cnt,biz_date,bug_cnt;<br>
  * 当前主键(包括多主键):<br>
  *	id;<br>
@@ -112,16 +111,16 @@ public class XmMenuState  implements java.io.Serializable {
 	BigDecimal planWorkerCnt;
 	
 	@ApiModelProperty(notes="总关闭bugs",allowEmptyValue=true,example="",allowableValues="")
-	Integer closedBugs;
+	BigDecimal closedBugs;
 	
 	@ApiModelProperty(notes="激活bugs",allowEmptyValue=true,example="",allowableValues="")
-	Integer activeBugs;
+	BigDecimal activeBugs;
 	
 	@ApiModelProperty(notes="已确认bugs总数",allowEmptyValue=true,example="",allowableValues="")
-	Integer confirmedBugs;
+	BigDecimal confirmedBugs;
 	
 	@ApiModelProperty(notes="已解决bugs总数",allowEmptyValue=true,example="",allowableValues="")
-	Integer resolvedBugs;
+	BigDecimal resolvedBugs;
 	
 	@ApiModelProperty(notes="产品编号",allowEmptyValue=true,example="",allowableValues="")
 	String productId;
@@ -351,25 +350,25 @@ public class XmMenuState  implements java.io.Serializable {
 	/**
 	 * 总关闭bugs
 	 **/
-	public void setClosedBugs(Integer closedBugs) {
+	public void setClosedBugs(BigDecimal closedBugs) {
 		this.closedBugs = closedBugs;
 	}
 	/**
 	 * 激活bugs
 	 **/
-	public void setActiveBugs(Integer activeBugs) {
+	public void setActiveBugs(BigDecimal activeBugs) {
 		this.activeBugs = activeBugs;
 	}
 	/**
 	 * 已确认bugs总数
 	 **/
-	public void setConfirmedBugs(Integer confirmedBugs) {
+	public void setConfirmedBugs(BigDecimal confirmedBugs) {
 		this.confirmedBugs = confirmedBugs;
 	}
 	/**
 	 * 已解决bugs总数
 	 **/
-	public void setResolvedBugs(Integer resolvedBugs) {
+	public void setResolvedBugs(BigDecimal resolvedBugs) {
 		this.resolvedBugs = resolvedBugs;
 	}
 	/**
@@ -628,25 +627,25 @@ public class XmMenuState  implements java.io.Serializable {
 	/**
 	 * 总关闭bugs
 	 **/
-	public Integer getClosedBugs() {
+	public BigDecimal getClosedBugs() {
 		return this.closedBugs;
 	}
 	/**
 	 * 激活bugs
 	 **/
-	public Integer getActiveBugs() {
+	public BigDecimal getActiveBugs() {
 		return this.activeBugs;
 	}
 	/**
 	 * 已确认bugs总数
 	 **/
-	public Integer getConfirmedBugs() {
+	public BigDecimal getConfirmedBugs() {
 		return this.confirmedBugs;
 	}
 	/**
 	 * 已解决bugs总数
 	 **/
-	public Integer getResolvedBugs() {
+	public BigDecimal getResolvedBugs() {
 		return this.resolvedBugs;
 	}
 	/**
